+++ NEWS.md
@@ -1,3 +1,21 @@
+# nanonext 0.8.1
+
+#### New Features
+
+* Implements synchronisation primitives from the NNG library. Condition variables allow the R execution thread to wait until it is signalled by an incoming message or pipe event.
+  + adds core functions `cv()`, `wait()`, `until()`, `cv_value()`, and `cv_reset()`.
+  + adds signalling receive functions `recv_aio_signal()` and `request_signal()`.
+  + `pipe_notify()` signals up to 2 condition variables whenever pipes are added or removed at a socket.
+* Adds `msg_pipe()` to return the pipe connection associated with a 'recvAio' message.
+* Exposes the `sha1()` cryptographic hash and HMAC generation function from the 'Mbed TLS' library (for secure applications, use one of the SHA-2 algorithms instead).
+* Utility function `'weakref<-'()` exposes `R_MakeWeakRef` from R's C API. Useful for keeping objects alive for as long as required by a dependent object.
+
+#### Updates
+
+* `ncurl_session()` gains a 'timeout' argument, and returns an 'errorValue' with warning upon error.
+* `listen()` and `dial()` gain the new logical argument 'error' to govern the function behaviour upon error.
+* Internal performance enhancements.
+
 # nanonext 0.8.0
 
 #### New Features